Refactor: Shuffled files, code and formatting
- Moved scenes in to to scenes folder
- Move Level related file in to level folder, including:
-- renaming files
-- re-grouping code
- Restyled to { after instead of on a newline
