--- src/level-definition.ts
@@ -1,20 +1,0 @@
-export let ElementCode = {
-    Solid: '0',
-    Passthrough: '1',
-    Pad: '2',
-    Ball: '3',
-    Bonus: '4',
-    Tile: '5'
-  }
-  
-export let  ElementCodeSolid = {
-    Brick: '0',
-    Pid: '1',
-    Spacebox: '2'
-  }
-  
-export let  ElementTile = {
-    Horizontal: '0',
-    Vertical: '1'
-  }
-
